Use named arguments for applying traits
Instead of introducing a dedicated syntax this can just use named
arguments. These arguments would be catched using a named splat argument
(e.g. "**aliases"). This basically means the signature of "use" would be
as following:

    def use(trait: Trait, **aliases: String) {
        ...
    }

Here "aliases" would be a Hash with the keys set to the variable names
and the values, well, the values.