Started moving back to a prototype based VM
The distinction between a Class and Object on VM level is very confusing
when these are not set up entirely the same way in the language. For
example, the "Class" struct would contain methods, constants, etc.
However, this struct would be exposed to the language by allocating an
Object of which the value would be ObjectValue::Class(...). This would
mean at least 2 structs would have to be allocated for every class
created. It's also confusing due to Object being used for both classes
and instances in the language while there's still a Class struct in the
VM.

After giving it more thought there _are_ ways to write a decent
prototype based VM that can be used for class based OO languages (or
really any other type of language), this commit contains a bulk of the
work needed to support that.

By using a prototype VM I don't have to deal with classes vs objects.
Instead there are only objects which can optionally have a prototype,
which really is just a parent object certain data is inherited from.

As an example, take the following Aeon code:

    class Person {
        def construct(name) {
            @name = name
        }
    }

This would result in the following instructions/compiled code objects:

    <main>
    ------

    string_literals
      0: "Person"
      1: "Object"
      2: "instance_prototype"

    code_objects
      0: CompiledCode(name="Person")

    0: get_toplevel 0
    1: get_const    1, 0, 1       # Object
    2: get_attr     2, 1, 2       # Object.instance_prototype
    3: set_object   2, 1          # Person
    4: set_name     2, 0
    5: set_object   3, 2          # Person.instance_prototype = ...
    6: run_code     4, 0, 2, 2, 3
    7: set_const    1, 2, 0

    class:Person
    ------------

    string_literals
      0: "construct"

    code_objects
      0: CompiledCode(name="construct")

    0: get_local  0, 0    # Person
    1: get_local  1, 1    # Person.instance_prototype
    2: def_method 1, 0, 0
    3: return     0

    method:construct
    ----------------

    string_literals
      0: "name"

    0: get_local 0, 0    # self
    1: get_local 1, 1    # name
    2: set_attr  0, 1, 0

To handle the difference between class and instance methods each "class"
keyword creates an object with an "instance_prototype" attribute. This
attribute contains an object on which methods are defined. When a class
creates a new instance of itself it will simply create a new object and
set its prototype to this "instance_prototype" attribute.

In this particular setup supporting singleton classes (or class methods
in general) is also super easy: they're just defined on the regular
object instead of the "instance_prototype" object.

Last, bootstrapping of core types has been removed from the VM in its
previous hardcoded form. Instead there's only one core object provided:
the top-level object. This object is required as looking up/setting
constants requires a source object.

Core types such as integer, float, etc, can be set up using the
following instructions:

* set_integer_prototype
* set_float_prototype
* set_string_prototype
* set_array_prototype

These simply take an object and will use this as the prototype for a new
object of one of these types. This means languages are free to choose
the names of these objects, their inheritance hierarchies (if any) and
so on. For example:

    string_literals
      0: "MyInteger"

    0: set_object            0
    1: set_name              0, 0
    2: set_integer_prototype 0

If the setting up of these types was hardcoded in the VM one would be
forced to using the hardcoded names (e.g. "Integer"), but in this case
one can set their own name.