Use shared heaps instead of thread-local heaps
Due to the possibility of cross-thread references a thread-local GC
would be stop-the-thread at the very least. It would have to either run
in-thread, meaning it would block the thread, or outside of it. The
latter means picking one of two options:

1. One GC thread per regular thread
2. One GC thread for all threads

Neither are ideal. The first one means 2x the amount of threads, the
second one means that the GC work queue could get clogged depending on
the amount of threads that are allocating objects.

Another problem is sharing objects between threads. To do so would
require specific instructions/metadata to mark certain objects as
shared. Instead of trying to provide memory safety on VM level this will
be the responsibility of the compiler.