Cleaned up Thread allocating and joining
Allocating threads no longer requires a prototype as doing so would
prevent the main thread from starting up.

Joining threads also works although the API is a bit weird. Calling
JoinHandle::join() consumes the JoinHandle, meaning that you have to
store it as an Option<JoinHandle> in a struct and call take() to take
over ownership. This in turn mutates the struct, requiring one to obtain
a write lock to the Thread struct. This is problematic because various
operations in a thread also require a write lock, leading to deadlocks
whenever one would try to join a thread.

To work around this the method Thread::take_join_handle() takes the
JoinHandle out of the struct and returns it. This still requires a write
lock but the lock can be dropped before actually joining the thread.
This means that to join a thread one has to do this:

    let handle = some_thread.write().unwrap().take_join_handle();

    if handle.is_some() {
        handle.unwrap().join();
    }

This is a bit ugly but prevents any deadlocks and removes the need for
faking a join using either channels or condition variables (the latter
would still likely still deadlock).

Currently there's some duplication when it comes to handling/printing
errors. While each thread has dedicated error handling code it's
unlikely for these errors to be printed unless one of those threads is
joined.

In following commits I'll most likely introduce a supervisor. Upon an
error a thread sends a message to this supervisor and exit. The
supervisor will then print the error, terminate all threads and exit. If
a thread itself were to try and terminate other threads it would need
some logic as to not terminate itself while in this process. By using a
separate supervisor thread we can work around this problem.