Break Object into Object and ObjectHeader
Instead of storing methods, attributes, constants and other information
directly in an Object this information is now stored in an ObjectHeader
struct. This struct is lazy allocated (using a Box) to cut down the size
of the Object struct. Prior to this change the Object struct was 320
bytes in size, whereas now it's "only" 104 bytes.