Use *const u8 in ObjectValue
Instead of ObjectValue defining enum variants with unique values
(String, isize, f64, etc) this enum now uses *const u8 for all variant
values. This keeps the enum more compact while still being able to check
what type of data we're dealing with.

The downside of this approach is the requirement of using unsafe code
and all the problems this might introduce.