Handle copying of local to global objects
Whenever a local object has to be stored in a global object it's first
copied to the global heap. For example, when setting a constant to a
local string the string object is copied to the global heap and the
global pointer is assigned to the constant. It's up to compilers to
ensure any instructions don't operate on the register containing the
local object after it has been copied to the global heap.