Fixed deadlock when shutting down threads
The deadlock would occur whenever a thread was signaled to stop while
also trying to remove itself from the thread list. Because the thread
that raised an error obtains a write lock on the thread list another
thread is not able to remove itself from said list (due to this also
needing a write lock).

To work around this only isolated threads are removed from the list
after successfully finishing their work. In case of an error related
shutdown there's no need to remove any threads as the system will be
shut down anyway.