Support for suspending processes
This refactors the scheduling of processes and how they're executed to
support suspending and resuming processes. Processes run for a certain
number of reductions (similar to Erlang) before they're suspended.
In the future the VM will also support automatically suspending when
running certain instructions.

With this commit the ability to manually start threads has been removed
as it's not very useful and leads to too much complexity.