Refactor Process to use interior mutability
This removes the need for locking the entire Process structure when
performing certain operations. When sending messages an object is now
allocated into a heap dedicated to messages instead of the eden heap.
This in turn removes the need for locking when allocating into the eden
heap.

The Inbox structure now uses a VecDeque instead of a regular Vec to
allow processing of messages in First-in First-out order instead of
First-in Last-out.