Make run_code less useless
This instruction now takes a runtime allocated Array for the arguments
making it much easier to use. There's also a new instruction called
"get_binding_of_caller" which returns the binding of a caller. This
instruction is primarily useful for closures so they can automatically
figure out what binding to use.