Refactor Object Pointers
This commit changes how object pointers are created and the data they
wrap. Instead of pointers being an enum to either a synchronized global
object or an unsynchronized local object a pointer is now just that: a
pointer to an object.

Currently object pointers are tagged (using pointer tagging) to mark
them as global. This may be removed in the future in favour of a system
where objects are considered to be global when their address resides in
the global heap address range (if this is possible in Rust).

The benefit of these changes is that pointers are now only 8 bytes and
allow for pointer tagging. Objects in turn are now only 32 bytes instead
of 48. This does require some unsafe code here and there but thankfully
this isn't too invasive.

Global pointers are no longer explicitly synchronized upon access using
an RwLock, or any other lock for that matter. Removing this greatly
simplifies the pointer setup and makes pointer tagging possible. On top
of that the locks would provide a false sense of safety. While the locks
would prevent concurrent writes they would not prevent cases where 1
thread would attempt to read data not yet written by another thread.
Operations such as defining methods and the likes can be synchronized in
a different way should this be deemed necessary in the future.