Continue work on Immix and refactor pointers
This continues work on the immix allocator by making it use the Rust
alloc/heap APIs. Object pointers have also been refactored to use the
TaggedPointer struct.

The pointer to an object header is now a tagged pointer containing extra
bits used to determine in what generation an object resides.