Use Immix for the global and mailbox heaps
The gloal heap is managed by the PermanentAllocator while the mailbox
heap is managed by MailboxAllocator. The Mailbox struct has been removed
in favour of keeping this code directly in Process as this makes it
possible to schedule GC runs and access a process' heap without having
to pass a Process to a Mailbox (which would create a circular
reference).

Currently the memory of mailboxes is not yet copied upon receiving. This
will be changed to also do a copy upon receive so that the young and
mature generations never refer to objects in the mailbox generation.
Garbage collection for the mailbox generation also isn't implemented
yet.