Use block allocation thresholds for scheduling GC
Instead of triggering a young collection every time a new block is
allocated the VM will now trigger a collection if a certain number of
blocks has been requested, regardless of them being re-used or
newly allocated blocks.

Currently the entire young generation has a single threshold. I may
introduce a separate threshold for the eden space to collect it more
frequently, but for now this will do.