Don't prepare eden when allocating LocalAllocator
When calling add_block the receiving Bucket must be in a stable memory
location. If we were to prepare the eden bucket we'd move it from the
stack to another location, invalidating any bucket pointers stored in
any allocated blocks.

The easiest way to work around this is to just not allocate an eden
block upon allocating a LocalAllocator. Doing so doesn't really have any
benefits anyway.