Implemented most of Immix
This includes evacuation, finalization of native resources, and a whole
bunch of changes to make all of this work. Evacuation is completely
untested and still has to be implemented for the mature generation.
Furthermore the code is a bit of a mess as I'm mostly comitting this so
I don't lose any changes.