Fix determining if evacuation is needed
Previously we'd iterate over individual blocks, check if they needed
evacuation, and if so update the "available" histogram. This could lead
to later code erroring out due to operating on a bin which contained
marked lines but no entry in the "available" histogram.

To handle this (and in a more correct way) we first iterate over all
blocks in a bucket to determine if evacuation is needed. If evacuation
is needed we then update the "available" histogram for all blocks, even
those that don't need evacuation. The GC also makes sure that blocks
without any holes are not evacuated as there's no point in doing so.