Potentially fix resolving of forwarding pointers
Previously forwarded pointers would be pushed into the objects to mark,
followed by a `continue` statement. For reasons not entirely clear to me
this would sometimes lead to forwarding pointers not being resolved
properly.

Since the pointer to forward is resolved when promoting/evacuating we
don't need the use of `continue`, instead we can just continue scanning
child pointers immediately. This appears to (thus far) solve the
problem.