Store recyclable blocks separately
The Immix paper states that when the allocator runs out of blocks it's
to _request_ a new recyclable block. Previously the collector would
"rewind" the allocator to the first block with a hole. This meant that
the allocator might have to skip over a large number of full blocks when
finding a new hole.

By storing recyclable blocks separately we can work around this. This
means that at the end of a collection cycle the regular list of blocks
only contains blocks that are either full or fragmented.

This does introduce some extra collector complexity which I hope to
clean up in following commits.