Further optimize garbage collection
This involves various micro optimizations as well as reworking how
objects are finalized.

To finalize objects we use a 3rd block level bitmap: the finalize
bitmap. Whenever an object that needs finalization is allocated it is
tracked in this bitmap. When an object is moved its entry in this bitmap
is removed. When a block is reclaimed during the coarse-grained sweep
the collector finalizes all unmarked objects in the "finalize" bitmap.