Initial implementation of the parallel GC
The GC can now process execution contexts (= stack frames) in parallel,
greatly increasing the collector throughput. Using benchmarks I measured
that processing frames in parallel instead of objects resulted in far
better performance. Most likely the large amount of objects would result
in a lot of synchronisation/locking overhead, whereas the amount of
stack frames is much smaller.

To achieve parallel collection the collector uses the "rayon" crate to
parallelize work. In early revisions I tried doing this manually using
the "crossbeam" crate but this lead to too much boilerplate code and
without good results.

To test the improvements I created a benchmark that creates a process
with 500 execution contexts. Each context contains 5000 strings in
registers, and 5000 strings in local variables (all unique objects).
This produces a total of 5 000 000 live objects. Marking these objects
takes around 46 milliseconds, producing a rate of 3.2 GB/sec. Since Aeon
processes have their own heap it's unlikely for them to ever reach 5
million live objects. As a result garbage collections should not take
more than 10-20 milliseconds in the best cases (though this is just an
estimate).

When objects have to be evacuated or promoted the throughput can drop
quite drastically depending on the amount of objects to move. For
example, when evacuating 980 objects the throughput drops to 2.9 GB/sec,
while promoting 5 million objects results in a throughput of 415 MB/sec.

The impact of moving objects can most likely be reduced somewhat by
changing the way objects are moved. Right now objects are stack
allocated, the ObjectValue is taken over, and this is then allocated.
Copying the memory directly may be faster.

== Parallel Moving

To support parallel moving of objects the collector acquires a spinlock
on the target Bucket. This allows objects in other buckets to be moved
around in parallel while still ensuring an object is only moved once.
This pattern is essentially as follows:

    if should_evacuate(pointer)
      lock(pointer.bucket)

      if should_evacuate(pointer)
        evacuate(pointer)
      else if should_resolve_forwarding_pointer(pointer)
        resolve(pointer)

      unlock(pointer.bucket)

This uses a spinlock (provided by the "parking_lot" crate) as Rust's
regular Mutex structures are really slow. This also relies on
parking_lot's nightly features to further reduce the overhead of these
spinlocks.

== Future Improvements

Preparing buckets, object finalisation, and reclaiming blocks still
happens in serial. Hopefully garbage collection timings can be further
reduced by doing this in parallel (e.g. per bucket or per block).