Fully implement process based garbage collection
As of this commit the Aeon GC is fully implemented. It supports heap
collections of a process as well as mailbox collections. Collection
statistics such as the number of marked objects, the throughput, and
other statistics are all tracked (though not yet exposed for
consumption).

Garbage collection is handled by two different collectors:

1. A parallel heap collector
2. A serial mailbox collector

The heap collector does most of its work in parallel, except for
processing the remembered set. The remembered set would have to grow
quite a bit before parallel collection would have any benefits.

The mailbox collector is a serial collector. Because mailboxes typically
don't contain that many pointers there's little benefit to performing
this work in parallel (at least for the time being).

== Evacuating Objects

In previous implementations it was possible for evacuated objects to
overwrite lines that were still in use. This was a result of line maps
being reset prior to collections (during the preparation phase). In an
Email conversation with the authors of Immix I was recommended to use a
line map that can essentially track the previous/current marks to work
around this problem. This is currently implemented for line maps only,
thus preventing evacuating of objects into used lines.

== Mailboxes

Mailboxes are essentially broken up into two regions: an external facing
region, and an internal region. When an external process sends a message
the message is stored in the external (and synchronised) region. If a
process sends a message to itself the message is stored in the
(unsynchronised) internal region.

When a process receives messages it will first consume the internal
region. This can be done without synchronisation. Once this region has
been consumed the process will acquire a lock on the external region,
copy all messages to the internal region, then proceed as usual.

When a process sends a message to itself the message is not copied. This
allows for efficient rescheduling of messages (e.g. when using something
like pattern matching to determine if you want to act upon a message).
These messages are stored in a separate list in the internal region,
allowing a heap collector to use this list as extra roots.

Fixes #51, fixes #53, fixes #54, fixes #57