Re-write logic for executing VM instructions
Instead of using a giant match statement the VM now uses a table of
function pointers. The numeric value of an instruction type is mapped to
an entry in this table. This reduces the overhead of instruction
execution while also cleaning up the code drastically.

This commit also introduces a set of tests for the various integer
instructions. Tests for the other instructions will be added in future
commits.

Fixes #63