Remove the notion of "self" in the VM
What "self" refers to is to be handled by a compiler, not the VM. An
easy way of doing this is to pass it as the first argument, but hide it
from the language. This simplifies the VM and allows for more flexible
code (e.g. "self" is no longer bound to a Binding).