Use Blocks for execution instead of CompiledCode
Instead of directly executing a CompiledCode using the run_code
instruction, the VM now operates on Blocks. These blocks wrap a
CompiledCode along with the binding of the scope the Block was created
in.

Using blocks removes the need for manually storing a binding in some
attribute and then having to look it up before executing a method.
Blocks also make it much easier to apply lexical scoping.