Rework suspending processes for garbage collection
Instead of a collector requesting a process to suspend itself, a process
now immediately suspends itself when garbage collection is necessary.
This greatly simplifies the code, and solves a race condition where
rescheduling a process could lead to it running during a collection
cycle; corrupting memory in the process.

A downside of this new approach is that a process may be suspended for a
longer period of time, but only if there are many processes scheduled
for collection and the collector threads are unable to keep up.