Use a Chunk for Register and Binding
A Chunk is simply a region of zeroed memory of a given type. It
essentially behaves like alloc::raw_vec::RawVec, but offers indexing
operations and zeroes memory. Zeroing is required as otherwise we might
operate on garbage memory, making it impossible to store a pointer in
the chunk and check if it's NULL or not.

Using a Chunk speeds up allocating registers and bindings as a Chunk
does not need to copy around an initial value for every index.