Revert "Use individual if statements for the dispatch loop"
This reverts commit 8d56273a9f596148ed833ad4e88409506b77073b.

The performance of if statements seems equally inconsistent compared to
a match statement. For example, removing SetInteger/Float/String in
favour of SetLiteral results in the match statement being faster
compared to the if statements.

For the time being we'll stick with a "match" and try to optimise this
as much as possible.