Use a single interpreter loop
Instead of using this:

    loop {
        while index < max {
            match instruction.instruction_type {
                ...
            }
        }
    }

We're now using this:

    loop {
        match instruction.instruction_type {
            ...
        }
    }

This removes the need for two loops (which seems to reduce the number of
branches quite a bit), and gives us some extra horizontal space. It does
require some trickery with macros and a little bit of unsafe code to
work around borrowing limitations, but overall it's not too bad.

Performance also appears to be better compared to the old approach, with
a boost of around 10% compared to the double loop setup.