Remove constants from the VM
There's no practical benefit to separating constants from attributes as
both are mutable on VM level. Further, the logic used for looking up
constants in the prototype chain is not necessary. Using module locals
and constants scoped to a module you rarely need to define a constant
inside another constant (e.g. a constant in a class). Removing all of
this logic simplifies the VM, compiler, and the runtime.