Remove syntax support for method requirements
While I really like the idea of requirements (or contracts) being part
of a method's signature there are several problems with this.

First of all it complicates the syntax. Method signatures can already
include a "throw" keyword to state what type will be thrown. This poses
the question: does "require" come before it, after it, or can you do
both? I chose to require "require" to be placed _after_ "throw" as this
makes parsing easier, but this means a developer has to keep in mind
that the order is throws -> requires, which is easy to forget.

Second is verbosity. Adding requirements to the type signature can lead
to it becoming very verbose. For example:

    fn [](index: Integer)
        throws Foo
        requires index > 0
        requires index < length
        -> T | Nil {
      ...
    }

This gets worse for complex requirements.

Finally, similar to regular assertions no sane error messages can be
generated other than something along the lines of "Method requirement
failed:" followed by the expression that failed. An example would be:

    Method requirement failed: index > 0

Such an error may be useful to the author of the code, but somebody
unfamiliar with the internals can get very confused. Adding custom error
messages would require even more syntax and would further increase the
verbosity of the signature.

Because of this I have decided to remove support for method contracts.
Instead Inko may one day provide something similar to assertions while
allowing custom error messages. This means it won't (unfortunately) be
part of the method signature, but at least it won't suffer from the
above problems.