Begin bootstrapping the type system
The way the type system will be bootstrapped is as follows: various raw
instructions will return pre-defined types, everything else will build
on top of this using runtime defined types.

In this commit we're adding the first steps towards this by providing a
GetBlockPrototype raw instruction which returns the built-in block
prototype type. The runtime will build upon this by using this
instruction in the right place when defining the Block class. If done
properly this should lead to the compiler picking things up nicely,
without having to be aware of specific class names (e.g. the integer
type being defined in "std::integer::Integer").