Embrace prototype OO and add more type info
The compiler now embraces prototype OO more, removing any mentions to
"classes", instead providing a simple "object" keyword. This keyword on
the outside behaves the same, but the underlying code/behaviour is much
more straightforward. For example, this:

    object Person {
      fn init(name: String) {
        let @name = name
      }
    }

Is (roughly) compiled down to this:

    let Person = Object.new

    fn(self) {
      fn new(name: String) {
        let instance = prototype.new

        instance.init(name)

        instance
      }

      fn init(name: String) {
        let @name = name
      }
    }.call(Person)

Here "Object.new" creates an empty object with its prototype set to
"Object". In this setup there are no class methods, but I actually quite
like this. Since you can define methods directly in a module there isn't
really a good reason to support class methods.

One big benefit of this approach is that bootstrapping is much simpler,
and the compiler can more easily infer types. I personally also find it
much easier to reason about objects when you have _just_ objects and
methods, instead of also having classes and meta classes.

With this also comes the decision to expose prototypes more to the
language. For example, Object will respond to the "prototype" message
instead of responding to some kind of "class" message.

Finally, in this commit the compiler is also changed to mostly rely on
message sending for defining core constructs. For example, when you
define a trait like so:

    trait ToString {
      fn to_string -> String
    }

It's (roughly) compiled down to:

    let ToString = Trait.new

    fn(self) {
      define_required_method('to_string')
    }.call(ToString)

Knowledge such as the type signatures of the required methods or any
type arguments will only be known to the compiler for the time being.