Start moving the compiler to Ruby
While Rust is a great language overall, I wasn't making as much progress
writing a compiler in it as I hoped for. In particular I found myself
fighting Rust more often than should be necessary. Perhaps I am to
blame, perhaps Rust was too strict.

Either way, since the goal is to make the compiler self hosting there's
no point in spending months trying to implement a Rust based compiler.
As such I decided to write the compiler in Ruby. Ruby has many flaws,
but it makes it much easier to quickly write a compiler without having
to fight the language all the time, at the cost of the occasional
runtime error during development. For example, porting over most of the
Rust code _and_ using a linear IR (instead of the tree IR used in the
Rust compiler) only took a few days.