Drop support for &&= and ||=
In Inko all binary operators are actually message sends. Message sends
in turn use eager evaluation for arguments, like most languages. This
however is something we don't want for binary AND and OR. For example,
consider the following example:

    fast_method || slow_method

This would translate to:

    fast_method.||(slow_method)

This means that `slow_method` would be invoked even if not necessary.

To work around this Inko will offer two separate messages: "and" and
"or". Both of these take a block they will only evaluate if necessary.
This means the above would be written as follows instead:

    fast_method.or { slow_method }

Because Inko will have these messages there's no need to also support &&
and ||, which also means &&= and ||= don't have to be supported.
Supporting the last two would also be difficult since you wouldn't be
able to use when with variables not defined using "let" or "var". That
is, in Ruby the following is valid:

    foo ||= 10

If `foo` is not defined it would be set to 10, otherwise it would be
left as-is. In Inko however you can only define a variable using "let"
or "var", and you can only reassign already existing variables. This
would then require syntax along the lines of this:

    let foo ||= 10

The complexity of supporting this does not justify the tiny benefit it
_may_ bring.

Using "&&" and "||" for messages is still valid syntax wise, Inko just
won't provide any methods for these messages.