Drop unions and casting for of optional types
Casting using the "as" keyword isn't going to work nicely in Inko since
there's no way for the compiler to know _how_ it has to cast an object.
Further, the "as" keyword was mostly added to allow casting of unions to
a specific member type.

Unions are removed because in Inko they are not terribly useful and you
can't really use them well without some kind of "switch", which I don't
want to introduce in Inko.

To deal with this Inko will instead provide "optional types", sometimes
called "nullable" types. The syntax for these types is as follows:

    object Array!(T) {
      fn [](index: Integer) -> ?T { ... }
    }

Here we define `[]` and state that the return type is either T or Nil.
This would be the most common use case for unions, but using optional
types makes things easier for the compiler and saves some typing.

Optional types are also valid for variable bindings:

    var number: ?Integer = 10
    number = Nil

They can also be used for method arguments:

    object Person {
      fn init(name: String, address: ?String) {
        let @name = name
        let @address = address
      }
    }