Hoist type, constant, and method definitions
Due to the compiler being a multi-pass compiler it's possible to use
certain types before they are defined. For example, the following used
to be valid:

    foo
    fn foo {}

This would be valid because the DefineTypes pass would define "foo",
which happens before we generate code in the ModuleBody pass. As a
result the ModuleBody pass would treat "foo" as defined on the first
line, even though it's not defined in the runtime until we reach the 2nd
line.

Hoisting methods, constants, and type definitions solves this problem by
moving these elements to the start of the scope. We _do not_ hoist local
variables or attributes, only the following is hoisted:

* Constants defined using "let" or "var"
* Objects defined using the "object" keyword
* Traits defined using the "trait" keyword
* Methods defined using the "fn" keyword

By hoisting we turn the above code into the following:

    fn foo {}
    foo