Support sending messages using module globals
This adds support for running methods defined as module globals. This is
done by storing the module itself in a global called "ThisModule". This
global is then used as the receiver when calling a method defined as a
global. This means that this:

    fn example { }

    object Foo {
      fn test {
        example
      }
    }

Is the exact same as this:

    fn example { }

    object Foo {
      fn test {
        ThisModule.example
      }
    }

With this change we also stop storing methods as globals since there's
no real benefit to it.