Removed throw reasons from the VM
Throw "reasons" were added to distinguish between errors being throws,
and values being thrown to break out of loops. This concept was taken
from Ruby, which implements "break" and "next" ("continue" in many
languages) using this system.

Instead of copying this, Inko will use tail recursion instead. This is a
bit more tricky to implement as tail-call optimisations have to be
added, but the resulting code will (in my opinion) be much cleaner.

As an example, a hypothetical infinite loop would be implemented as
follows:

    var number = 0

    {
      (number < 10).if true: {
        number += 1
        True
      }, false: {
        False
      }
    }.while_true

With the implementation of "while_true" being along the lines of:

    fn while_true -> Nil {
      call.if_false { return }
      while_true
    }