Removed SendMessage instruction
This instruction was introduced in the past to combine GetAttribute and
RunBlock, thus increasing performance a little bit. Unfortunately having
a special instruction like this makes it harder to introduce other
specialised instructions. For example, introducing a separate
instruction for executing a block with a rest argument would require
duplicating this logic for both RunBlock and SendMessage.

Another problem is that SendMessage would panic when sending a message
not understood by an object. On the other hand, when using GetAttribute
and RunBlock separately the compiler would generate the right code to
delegate such message sends to a different method.

In short, removing this instruction _may_ slow things down a tiny bit
but it also gives the compiler a lot more flexibility.