Added proper support for implementing traits
This is done using a few separate bits. First we can now use any pointer
for attribute keys, as long as the pointers used for lookups and writes
are identical. This means you can't use process-local objects reliably,
but it's good enough when using permanent objects.

Next we implement `Object.implement_trait` in such a way that it used
the above feature to efficiently track which traits are implemented by
an object. This makes it easy/fast to check for this in
`Object.kind_of?`, without having to mess with prototype chains or
having to create a dedicated Set data type.

All of this means we can now run code such as the following:

    trait Thing {}
    object Person {}
    impl Thing for Person {}

    Person.new.kind_of?(Thing) # => True