Use the system allocator by default
While jemalloc may provide better performance it also comes with much
more memory usage. For example, a basic hello world may use a few MB of
memory when using the system allocator. On the other hand with jemalloc
it can easily use 40 MB. For most users such an increase of memory is
probably not acceptable, so we'll switch to the default allocator by
default.

Rust is planning to also switch, and once done we can get rid of the
Cargo feature and related code.