Generate code for handling unknown messages
The compiler will now generate the right code so the runtime can handle
objects being sent messages for which they don't have any corresponding
methods. Keyword arguments that are not mapped directly to positional
arguments by the compiler are currently ignored.

With this in place we can implement the necessary bits and pieces for
Nil to work properly. In particular Nil now behaves in such a way that
sending an unknown message to it will produce another Nil.